+++ src/st-table/index.ts
@@ -1,3 +1,0 @@
-export * from './shared/st-list.model';
-export { Order, ORDER_TYPE, Page, DataList, Model } from './shared';
-export * from './st-list.module';