+++ src/st-table/shared/index.ts
@@ -1,6 +1,0 @@
-export * from './st-list-body';
-export * from './st-list-header';
-export * from './st-list-row';
-export * from './st-list.model';
-export * from './st-pagination';
-