+++ src/st-table/shared/st-list-body/index.ts
@@ -1,1 +1,0 @@
-export * from './st-list-body.component';