+++ src/st-table/shared/st-list-facelet/index.ts
@@ -1,1 +1,0 @@
-export {FaceletBlock, FaceletCriteria, Filter, FILTER_OPERATOR} from './st-facelet.model';