+++ src/st-table/shared/st-list-header/index.ts
@@ -1,1 +1,0 @@
-export * from './gosec-list-header.component';