+++ src/st-table/shared/st-pagination/st-pagination.model.ts
@@ -1,7 +1,0 @@
-export class Page {
-  constructor(
-    public currentPage: number,
-    public elementsPerPage: number,
-    public limit: number // TODO: Name more descriptive
-  ) { }
-}