+++ src/st-table/shared/st-search/gosec-search.model.ts
@@ -1,8 +1,0 @@
-export interface SearchElement {
-  id: string;
-  label: string;
-}
-
-export class Search {
-  constructor(public searchBy: string, public searchText: string) {}
-}